refactor(domain): removed package model and moved content up
