+++ kingdom_sdk/domain/aggregate.py
@@ -1,8 +1,8 @@
 from abc import ABC
 from typing import Iterable
 
-from kingdom_sdk.domain.model.entity import Entity
-from kingdom_sdk.domain.model.event import Event
+from kingdom_sdk.domain.entity import Entity
+from kingdom_sdk.domain.event import Event
 
 
 class Aggregate(Entity, ABC):
