+++ kingdom_sdk/domain/command.py
@@ -1,0 +1,5 @@
+from kingdom_sdk.domain.message import Message
+
+
+class Command(Message):
+    pass
