+++ kingdom_sdk/domain/event.py
@@ -1,0 +1,5 @@
+from kingdom_sdk.domain.message import Message
+
+
+class Event(Message):
+    pass
