--- kingdom_sdk/domain/model/command.py
@@ -1,5 +1,0 @@
-from kingdom_sdk.domain.model.message import Message
-
-
-class Command(Message):
-    pass
