--- kingdom_sdk/domain/model/event.py
@@ -1,5 +1,0 @@
-from kingdom_sdk.domain.model.message import Message
-
-
-class Event(Message):
-    pass
