+++ kingdom_sdk/domain/value_object.py
